Move pages to separate files
