+++ lib/pages/auth_page.dart
@@ -1,0 +1,21 @@
+import 'package:flutter/material.dart';
+import 'package:testproject/authentication.dart';
+
+class AuthenticationPage extends StatefulWidget {
+  const AuthenticationPage({Key? key}) : super(key: key);
+
+  @override
+  State<AuthenticationPage> createState() => _AuthenticationPageState();
+}
+
+class _AuthenticationPageState extends State<AuthenticationPage> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text("TODO List App"),
+      ),
+      body: const Authentication(),
+    );
+  }
+}
