+++ lib/pages/signup_page.dart
@@ -1,0 +1,21 @@
+import 'package:flutter/material.dart';
+import 'package:testproject/signup.dart';
+
+class SignupPage extends StatefulWidget {
+  const SignupPage({Key? key}) : super(key: key);
+
+  @override
+  State<SignupPage> createState() => _SignupPageState();
+}
+
+class _SignupPageState extends State<SignupPage> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text("TODO List App"),
+      ),
+      body: const Signup(),
+    );
+  }
+}
