+++ lib/pages/todo_page.dart
@@ -1,0 +1,21 @@
+import 'package:flutter/material.dart';
+import 'package:testproject/todo.dart';
+
+class TodoPage extends StatefulWidget {
+  const TodoPage({Key? key}) : super(key: key);
+
+  @override
+  State<TodoPage> createState() => _TodoPageState();
+}
+
+class _TodoPageState extends State<TodoPage> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text("TODO List App"),
+      ),
+      body: const Todo(),
+    );
+  }
+}
